requirements.txt
# Vibe Ideation Validator - Dependencies
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0
